+++ storybook/src/components/accordion/accordion.content.ts
@@ -71,6 +71,19 @@ function section4({ richContentTemplate }: Templates): AccordionSection<Template
   };
 }
 
+const sectionWithHeadings: AccordionSection<TemplateResult> = {
+  handleTitle: "Inhoudsopgave",
+  heading: "h2",
+  open: true,
+  content: html`<h2>Aanvullende regels</h2>
+    <h3>Uit andere besluiten</h3>
+    <h4>Uit hoofdregeling omgevingsplan</h4>`,
+};
+
+export function compactSections(templates: Templates): AccordionSection<TemplateResult>[] {
+  return [...basicSections(templates), sectionWithHeadings];
+}
+
 export function basicSections(templates: Templates): AccordionSection<TemplateResult>[] {
   return [
     section1,